requirements.txt
torch
torchvision
transformers
pillow
numpy
opencv-python
matplotlib
streamlit
